Added Swap On 7 functionality
Added a perform action check that checks for the 7 card, if the option is enabled it will go to the SwapHandsOnSeven function, there it will either prompt the user to select another player that is valid, ie not the player and not a player that isn't playing. The computer will loop through and find the player with the lowest card count to swap. The PlayerSwapHands function then gets called to perform the swap by storing the references to the players hands and then reassigning them.